+++ src/boto3.ts
@@ -1,9 +1,9 @@
-import { window, Progress } from 'vscode';
+import { window, Progress, ExtensionContext } from 'vscode';
 import { createSmartInstaller } from './installers/smart';
 
 
-export async function getOrInstallBoto3Version(): Promise<string> {
-    const smartInstaller = await createSmartInstaller();
+export async function getOrInstallBoto3Version(context: ExtensionContext): Promise<string> {
+    const smartInstaller = await createSmartInstaller(context);
     const boto3Version = await smartInstaller.getBoto3Version();
     if (boto3Version) {
         return boto3Version;
@@ -17,17 +17,4 @@ export async function getOrInstallBoto3Version(): Promise<string> {
 
     await smartInstaller.install("boto3", "", false);
     return await smartInstaller.getBoto3Version();
-}
-
-
-export async function updateBoto3Version(version: string): Promise<string> {
-    const smartInstaller = await createSmartInstaller();
-    const action = await window.showInformationMessage(
-        `New boto3 version ${version} available!`,
-        `Update with ${smartInstaller.getInstallers().map(x => x.name).join(' / ')}`,
-    );
-    if (!action) { return ""; }
-
-    await smartInstaller.install("boto3", version, false);
-    return version;
 }