+++ src/commands/quickstart.ts
@@ -1,16 +1,16 @@
 import { getServicePackages } from "../utils";
-import { workspace, Uri, window, env } from "vscode";
+import { workspace, Uri, window, env, ExtensionContext } from "vscode";
 import { getOrInstallBoto3Version } from '../boto3';
 import modifyHandler from './modify';
 import updateHandler from './update';
 import autodiscoverhandler from './autodiscover';
 
 
-export default async function handle(): Promise<void> {
-    const boto3Version = await getOrInstallBoto3Version();
+export default async function handle(context: ExtensionContext): Promise<void> {
+    const boto3Version = await getOrInstallBoto3Version(context);
     if (!boto3Version) { return; }
 
-    const servicePackages = await getServicePackages();
+    const servicePackages = await getServicePackages(context);
     const pylanceEnabled = workspace.getConfiguration('python').get('languageServer') === 'Pylance';
     const typeCheckingEnabled = workspace.getConfiguration('python').get('analysis.typeCheckingMode') !== 'off';
     const autoCompleteEnabled = servicePackages.filter(x => x.installed && x.getExtraName()).length > 0;
@@ -28,12 +28,12 @@ export default async function handle(): Promise<void> {
         env.openExternal(Uri.parse('https://marketplace.visualstudio.com/items?itemName=ms-python.vscode-pylance'));
     }
     if (action === 'Modify services') {
-        await modifyHandler();
+        await modifyHandler(context);
     }
     if (action === 'Auto-discover services') {
-        await autodiscoverhandler();
+        await autodiscoverhandler(context);
     }
     if (action === 'Update') {
-        await updateHandler();
+        await updateHandler(context);
     }
 }