+++ src/extension.ts
@@ -4,25 +4,41 @@ import modifyHandler from './commands/modify';
 import docsHandler from './commands/docs';
 import autodiscoverHandler from './commands/autodiscover';
 import quickstartHandler from './commands/quickstart';
+import installerHandler from './commands/installer';
 import { createSmartInstaller } from './installers/smart';
 
 export async function activate(context: ExtensionContext) {
     console.log('boto3-ide is now active!');
 
     context.subscriptions.push(
-        commands.registerCommand('boto3-ide.quickstart', quickstartHandler)
+        commands.registerCommand('boto3-ide.quickstart', async () => {
+            return await quickstartHandler(context);
+        })
     );
     context.subscriptions.push(
-        commands.registerCommand('boto3-ide.modify', modifyHandler)
+        commands.registerCommand('boto3-ide.modify', async () => {
+            return await modifyHandler(context);
+        })
     );
     context.subscriptions.push(
-        commands.registerCommand('boto3-ide.update', updateHandler)
+        commands.registerCommand('boto3-ide.update', async () => {
+            return await updateHandler(context);
+        })
     );
     context.subscriptions.push(
-        commands.registerCommand('boto3-ide.docs', docsHandler)
+        commands.registerCommand('boto3-ide.docs', async () => {
+            return await docsHandler(context);
+        })
     );
     context.subscriptions.push(
-        commands.registerCommand('boto3-ide.autodiscover', autodiscoverHandler)
+        commands.registerCommand('boto3-ide.autodiscover', async () => {
+            return await autodiscoverHandler(context);
+        })
+    );
+    context.subscriptions.push(
+        commands.registerCommand('boto3-ide.installer', async () => {
+            return await installerHandler(context);
+        })
     );
 
     // await hello(context);
@@ -33,9 +49,9 @@ async function hello(context: ExtensionContext): Promise<void> {
     console.log("initialized", initialized);
     if (!initialized) {
         context.workspaceState.update("initialized", "true");
-        const smartInstaller = await createSmartInstaller();
+        const smartInstaller = await createSmartInstaller(context);
         if (await smartInstaller.getBoto3Version()) {
-            await quickstartHandler();
+            await quickstartHandler(context);
         }
     }
 }