+++ src/utils.ts
@@ -1,4 +1,4 @@
-import { window, ProgressLocation, Progress, QuickPickItem } from 'vscode';
+import { window, ProgressLocation, Progress, QuickPickItem, ExtensionContext } from 'vscode';
 import { servicePackages } from './servicePackages';
 import { Boto3StubsPackage, PypiPackage } from './pypi';
 import { createSmartInstaller } from './installers/smart';
@@ -15,8 +15,8 @@ export async function showProgress(message: string, progressCallback: (progress:
 }
 
 
-export async function getServicePackages(recommended: string[] = []): Promise<PypiPackage[]> {
-    const smartInstaller = await createSmartInstaller();
+export async function getServicePackages(context: ExtensionContext, recommended: string[] = []): Promise<PypiPackage[]> {
+    const smartInstaller = await createSmartInstaller(context);
     const boto3Version = await smartInstaller.getBoto3Version();
     const installedPackages = await smartInstaller.pip.listPackages();